requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.0.11
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==42.0.5
cssselect2==0.7.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.4
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-heroku==0.3.1
django-tailwind==3.8.0
django-tinymce==4.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dm-tree==0.1.8
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.1
filelock==3.13.3
Flask==3.0.2
flatbuffers==24.3.7
fqdn==1.5.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.3
ipykernel==6.29.4
ipython==8.23.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keras==3.0.5
libclang==16.0.6
lxml==5.2.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
opt-einsum==3.3.0
oscrypto==1.3.0
overrides==7.7.0
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pillow==10.2.0
platformdirs==2.6.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyPDF3==1.0.6
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-json-logger==2.0.7
pytz==2021.3
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
referencing==0.34.0
reportlab==3.6.13
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
Send2Trash==1.8.3
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
sslcommerz-lib==1.0
stack-data==0.6.3
starlette==0.36.3
svglib==1.5.1
tailwind==3.1.5b0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
terminado==0.18.1
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
uritools==4.0.2
urllib3==1.26.18
virtualenv==20.14.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.41.2
whitenoise==6.2.0
widgetsnbextension==3.5.2
wrapt==1.16.0
xhtml2pdf==0.2.8
